Improve the efficiency of searches that return huge result sets.
The core search returns *all* result objects, even though only a small subset is usually displayed on a page at one time. The result set is counted to provide 'hits' information and to construct the pagination controls. However, if the result set is large this is very resource intensive and can crash out.

It is more efficient to conduct the search in two stages: A simple getCount() to determine how many results there are, followed by getObjects using the offset and limit to return only those results that are actually needed for display. The 'hits' information can be preserved by padding the results array out to the same length, using simple TRUE values (still not ideal, but still saves a ton of resouces).

git-svn-id: https://subversion.assembla.com/svn/impresscmsaddons/modules/library_ipf/trunk@24871 7a4d4d67-0c3a-0410-b24c-89d50c169223